Add early fusion transformer
