+++ aiproteomics/frag/models/__init__.py
@@ -1,1 +1,1 @@
-from .transformer_frag import *
+from .early_fusion_transformer import *
